view models bugfixing

+++ Core/ViewModels/ProductViewModel.cs
@@ -7,7 +7,7 @@ namespace Core.ViewModels
 {
     public class ProductViewModel
     {
-        public int? id { get; set; }
+        public long? id { get; set; }
         public string name { get; set; }
         public string description { get; set; }
         public string description_short { get; set; }
